requirements.txt
astroid==2.0.4
autopep8==1.4.2
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
coveralls==1.5.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.2
django-cors-middleware==1.3.1
django-extensions==2.1.3
django-filter==2.0.0
django-heroku==0.3.1
django-nose==1.4.6
django-rest-passwordreset==0.9.7
django-rest-swagger==2.2.0
django-taggit==0.23.0
django-taggit-serializer==0.1.7
djangorestframework==3.9.0
docopt==0.6.2
drf-yasg==1.11.0
Faker==1.0.0
flake8==3.6.0
gunicorn==19.9.0
idna==2.7
inflection==0.3.1
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
nose==1.3.7
oauthlib==2.1.0
openapi-codec==1.3.2
pep8==1.7.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycodestyle==2.4.0
pyflakes==2.0.0
PyJWT==1.6.4
pylint==2.1.1
python-dateutil==2.7.5
python-decouple==3.1
python-http-client==3.1.0
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
requests==2.20.0
requests-oauthlib==1.0.0
ruamel.yaml==0.15.75
sendgrid==5.6.0
simplejson==3.16.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.24
whitenoise==4.1
wrapt==1.10.11
